+++ app/auth/auth.py
@@ -3,7 +3,7 @@ from sqlalchemy import select
 from ..config import config
 from fastapi import Depends, Request
 from fastapi_users import BaseUserManager, FastAPIUsers
-from app.dal.user import UserRepository
+from app.repositories.user import UserRepository
 
 from app.models.models import User
 from app.db import get_user_db
