+++ app/db.py
@@ -2,7 +2,7 @@ from collections.abc import AsyncGenerator
 from .config import config
 
 from fastapi import Depends
-from app.dal import UserRepository, TaskRepository
+from app.repositories import UserRepository, TaskRepository
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
 from .models.models import Task, User, Base
